* Moved destination path logic from form-field-model to upload-cloud widget component
