+++ lib/core/form/components/widgets/core/form-field-file-source.ts
@@ -33,14 +33,3 @@ export interface FormFieldFileSource {
     serviceId: string;
     destinationFolderPath: DestinationFolderPath;
 }
-
-export enum FileSourceTypes {
-    ALL_FILE_SOURCES_SERVICE_ID = 'all-file-sources',
-    ALFRESCO_CONTENT_SOURCES_SERVICE_ID = 'alfresco-content'
-}
-
-export enum DestinationFolderPathType {
-    STATIC_TYPE = 'value',
-    STRING_TYPE = 'string',
-    FOLDER_TYPE = 'folder'
-}
