+++ lib/core/form/components/widgets/core/form-field-types.ts
@@ -55,8 +55,4 @@ export class FormFieldTypes {
     static isContainerType(type: string) {
         return type === FormFieldTypes.CONTAINER || type === FormFieldTypes.GROUP;
     }
-
-    static isUploadType(type: string) {
-        return type === FormFieldTypes.UPLOAD;
-    }
 }
